move pseudonym model from service to component
