+++ src/components/pseudonym/repo/index.ts
@@ -1,3 +1,4 @@
 import * as pseudonymRepo from './pseudonym.repo';
+import type { Pseudonym } from './db/pseudonym';
 
-export { pseudonymRepo };
+export { pseudonymRepo, Pseudonym };
